chore: move tests into __tests folder so they get run in CI
