+++ packages/wrangler/src/__tests__/pages/filepath-routing.test.ts
@@ -1,9 +1,12 @@
 import { mkdirSync, writeFileSync } from "fs";
 import { runInTempDir } from "../../__tests__/helpers/run-in-tmp";
+import {
+	compareRoutes,
+	generateConfigFromFileTree,
+} from "../../pages/functions/filepath-routing";
 import { toUrlPath } from "../../paths";
-import { compareRoutes, generateConfigFromFileTree } from "./filepath-routing";
+import type { HTTPMethod, RouteConfig } from "../../pages/functions/routes";
 import type { UrlPath } from "../../paths";
-import type { HTTPMethod, RouteConfig } from "./routes";
 
 describe("filepath-routing", () => {
 	describe("compareRoutes()", () => {
